Implement a "Thread Local Storage" component in Proc
For Linux support, we're going to have to have thread local storage (TLS).

Initially the way I implemented this was quite wrong. This change
sets it up in a reasonably portable way: add a TLS component to Proc,
and for the amd64 architecture, hook the arch_prctl system call up
to set it. The messiness of this will all be hidden in the linux
'return from syscall' code (patch coming).

Change-Id: I9231cad13b10e33eaf6a5e68dfd6a0675a37a125
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/54
Tested-by: Jenkins QA
Reviewed-by: John Floren <john@jfloren.net>