Remove a useless test for PCI config type which breaks PCI on qemu-system-x86_64
Plan 9 code tests for a type 1 or 2 config mechanism. Type 1 has been
implemented everywhere we've been, and type 2 is almost never
implemented; we've pulled the test from coreboot because it's useless.
All configs of type 1 will always work.

Still worse, the type 1/type 2 test in plan 9 fails on qemu-system-x86_64,
such that no pci devices are detected because no functioning config
mechanism is found.

We can do further simplification but the pcicfginit variable is
also used as a 'we have configured pci' flag, so that cleanup will
have to come later.

Change-Id: I8269b8d1221514f935b157caf1fa8213690ed8a4
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/61
Tested-by: Jenkins QA
Reviewed-by: Akshat Kumar <seed@mail.nanosouffle.net>