Vector unknown linux system calls to user mode
Linux programs are shepherded by the linux command.
In some cases we can let the linux command manage system calls we
don't want to put in the nxm kernel. Hence, when a system call is
in range but not supported, we post a note in the form
linux:%d %#ullx %#ullx %#ullx %#ullx %#ullx %#ullx
with the system call number and the (up to) 6 arguments.
The linux command can then decide to do something or
just bail out.

Change-Id: I6b013e7a4a86e663fc22146f280b780f37835289
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>