Add thread local storage system call for Go.
Add a settls system call, to support Go. We don't do this via a write to ctl
because that is s-l-o-w, and this needs to be fast; we may optimize for it
in assembly later if it is not good enough.

Move the wrmsr to kexit, and only do the wrmsr if up->tls is non-nil.
Be sure to clear tls in newproc().

Change-Id: If8b9f77bc8acfd608d6c733f8f0ec3436f1fa028
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/71
Tested-by: Jenkins QA
Reviewed-by: Akshat Kumar <seed@mail.nanosouffle.net>