Support user-level management of Linux system calls
We won't support all 300 linux system calls in the kernel.
This code allows user mode programs to catch them in notify, handle them,
and call noted. We can now completely intermix linux and NxM system calls.
The changes to l64syscall.s and linuxsyscall.c are a bit tricky, but
are tested and work well.

Change-Id: I8881520338d1d9f692494f70e487e5ba0372146a
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/76
Tested-by: Jenkins QA
Reviewed-by: John Floren <john@jfloren.net>