nxm: support for cputype and awk
The cputype environment variable is used in several
mkfiles, including awk's, and as-is, it does not make
much sense on UNIX. This is a proposal to reconcile
its usage with cross-builds: simply use

cputype=$cpu/unix

since it is mostly the toolchain information that is
desired by most mkfiles.

This perspective is used here in cross-compiling awk
with the most minimal changes to sys/src.

Change-Id: Iaf36114c4c0270aa000e340beb09e0ec0b23bfd4
Signed-off-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-on: http://nxm.coreboot.org/review/89
Tested-by: Jenkins QA
Reviewed-by: John Floren <john@jfloren.net>