amd64/mkfile: use the INCLUDE environment variable
The util/(6 c)^c toolchain affords us a special INCLUDE
environment variable, which specifies a default location
where 6c will look for include files, when it is not to be
concerned with the Plan 9 default locations. Since we are
cross-compiling, this is exactly our situation. Moreover,
when $NXM is nil, the CFLAGS options

-I $NXM/$objtype/include -I $NXM/sys/include

account for all of the Plan 9 defaults.

Thus, here we can replace -I $NXM with INCLUDE=$NXM.
When $NXM is nil (i.e., in NxM), the former makes
`6c $CFLAGS ...' invalid. On the other hand, the latter works
in any case.

Change-Id: I076346b29a44907479ecc561759f5058aecdc3c4
Signed-off-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-on: http://nxm.coreboot.org/review/91
Tested-by: Jenkins QA
Reviewed-by: John Floren <john@jfloren.net>