Clean up linux emulation
Add support for mmap/munmap (which does not work) just to show how
it might be done. I'm leaving it in the kernel now as I don't know WTF
is going on with it. But intermixing nxm calls to segbrk seems to interact
badly with everything else. Add trace.c from linuxemu3. Remove a
print that should not be used. Note that the trace dump can be called
AFTER the linux program has exited or if you don't intend to restart it.

Change-Id: Ib306fa396a355471819869f0dcb4a3ef5ea34a0a
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/93
Tested-by: Jenkins QA
Reviewed-by: John Floren <john@jfloren.net>