Better Rune support for APE and a floating point fix for 6l
The changes to APE are taken from LSUB¹ and Plan9Ports.
These are mostly required because we use, in the spirit of
Plan9Ports, Runemask = 0x10FFFF, and not 0xFFFF.
They are necessary for applications that depend on APE,
such as awk.

Awk was also the reason for the (minor) changes to
6c/6.out.h. These are required because the internal
IEEE representation of the double-precision data
type depends on two 32-bit components, whereas
we have `Ieee.h' and `Ieee.l' both declared as `long',
which on UNIX systems is 64-bit wide. Hence, any
code going through 6c with data of type `double'
would have both of its internal components padded
to 64-bits. When the object data comes around to
6l, it will manage the floating point data with the
proper MOVSD calls into the FPU registers. Since
these instructions only operate on 64-bit data, they
would only ever pick up the lower quadword (viz.,
`Ieee.l').

As an example: the number 0.0000004000000000125
has internal IEEE representation as:
	0x3e9ad7f2.0x9ac0499f
(hi.lo). Since the `lo' component is 64-bit wide on
UNIX, through a series of unfortunate events inside
6c, it would come out to be:
	0xffffffff9ac0499f,
thus making the full representation:
	0x000000003e9ad7f2.0xffffffff9ac0499f;
noting that the padding with zeroes cannot be
overlooked (if indeed lo = 0, then the whole double
precision number turns out to be 0). Of course,
this representation, when the lower order bits are
picked up, turns into:
	0xffffffff.0x9ac0499f,
which has the form of a NaN.

Since this is a delicate matter, highly dependent on
data type widths, I've proposed to include a regression
test for future purposes: float.c. For now, it simply
makes sure that the internal representation of a `double'
agrees with what is expected; and beyond that, it
requires that we can perform a simple division without
catching an exception. More tests, such as proper
NaN handling, should be added in the future.

¹ http://codereview.appspot.com/5683071/

Change-Id: I6c83a1520238632835d4831ea79097f1a327ecad
Signed-off-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-on: http://nxm.coreboot.org/review/97
Tested-by: Jenkins QA
Reviewed-by: ron minnich <rminnich@gmail.com>