linux: minor fixes in regress/badsyscall and cmd/linux
The `linux' shepherd sets up its note handler twice,
which is unnecessary.

Due to the noted(NDFLT) in badsyscall's handler,
the final test (for a bad syscall) causes the program
to go into a broken state, instead of coming back
and finishing up the test. It is better to make this a
clean finish.

Change-Id: I6e08b29398506f1fc684d32fb9a7efd35722ddb5
Signed-off-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-on: http://nxm.coreboot.org/review/100
Tested-by: Jenkins QA
Reviewed-by: ron minnich <rminnich@gmail.com>