src/cmd: add support for a slew of programs and libs
We can compile many programs now, so no need
to shun them. This also led to a couple of trivial
fixes in the toolchain:

in 6l, we meant to search for libraries in "." first, but
were never doing that; this change adds dot to the
beginning of the directory search array. It also
cleans up path names a wee bit and adds a debug
option to view how libraries get searched.

In 6c, the path for `cpp' was hard-coded to /bin,
which for us is not true; just use relative paths instead
(exec will handle them).

I've pulled cmd/acme and sys/lib/scsicodes directly
from nix. The latter is needed to build cmd/usb.

Finally, I've updated .gitignore to reflect all of these
changes (e.g., eqn.c was mk-generated).

Change-Id: I4c54388129e001732d831cb84397b554663374b3
Signed-off-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-on: http://nxm.coreboot.org/review/119
Tested-by: Jenkins QA
Reviewed-by: John Floren <john@jfloren.net>