nxm/util: add wrkey, fix kernel mkfile to produce nvram, etc.
We can now produce the nvram during kernel compile time, from
Unix, using wrkey(1). (As noted in the k10/mkfile, there is a BUG:
Jenkins cannot yet provide input to wrkey, so it fails to build. When
this is fixed, the mkfile will again be updated.)

Also a slight bit of tidying-up: removed src/nxm/root/nvram
and updated .gitignore so that our passwords are not tracked;
updated the util/BUILD script; removed the unnecessary
util/readnvram.c source, since we just link against libauth[srv]
from Plan 9 Ports.

Also added xbin/wrkey to .gitignore so that the binary is not
tracked.

Change-Id: Ic588268c4f55a290f032fa02e94597ec70b0a64f
Signed-off-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-on: http://nxm.coreboot.org/review/124
Tested-by: Jenkins QA
Reviewed-by: ron minnich <rminnich@gmail.com>