Fix TLS and getpgszi
There was lots of confusion in the getpgszi code. getpgszi was assumed to
return -1 in some places, always valid values in others.

It now returns a valid value. We now check that it returns a certain value
in the one place it matters.

TLS should always be set, even when zero. And a process should be allowed to
set it to 0, which is an invalid address. Delete the test.
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

Change-Id: Icca5ea37d644bff36e03833b07a7527707f8c239
Reviewed-on: http://nxm.coreboot.org/review/137
Tested-by: Jenkins QA
Reviewed-by: John Floren <john@jfloren.net>