src/nxm: make #c/sysstat follow 9P specs regarding read()
We were returning the whole file upon any read. This would
cause buffer overflows when the expected length of the read
is less than the size of the whole file. An example of this case
occurs in getproccount() in the Go runtime, with a large
number of CPUs.

The optimization is still to use a large buffer size for frequent
reads on the sysstat file, but at least now the underlying
behaviour is not broken.

A regression test has been added to cmd/regress/sysstatread.c
to document the issue.

Change-Id: I1cfd642d29a3955499b3507f56f0449d1710700e
Signed-off-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-on: http://nxm.coreboot.org/review/153
Tested-by: Jenkins QA
Reviewed-by: ron minnich <rminnich@gmail.com>