all: restructure tos; include PID in the default process TLS
The discrepancy in the size of `long' between Plan 9 and
UNIX causes miscalculation of the offset of `pid' from Tos,
in programs that must explicitly calculate that, such as
the Go runtime. Thus, we include it in the default TLS for
all processes so that the Go runtime can access it that
way, and move it to the "top of stack" so as to resolve
the offset miscalculation problem in the future (and also
to preserve the place of the Go pointers in TLS, for the Go
runtime). Moreover, we explicitly make it `uvlong' so as
to support 64-bit PIDs.

Change-Id: I810048be1586f758b864c968580045d9bcb1fa11
Signed-off-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-on: http://nxm.coreboot.org/review/156
Tested-by: Jenkins QA
Reviewed-by: ron minnich <rminnich@gmail.com>