Add a type of identity-mapped segment.
If syssegbrk is called with a base of 1024*GiB, the page will be
allocated in thg SG_PV segment. In this segment, the address
returned is a direct map to the page phsysical address. This
will in turn allow, among other things, interrupt handlers to directly
stage IO to process memory, assuming other things are done
correctly (as we did on BG/P).

The next interesting step would be to allow a process to loan
out a page in the SG_PV space via a write to ctl file. Then
processes would have a space where all pointers are valid, which
might be very useful.

This is an experiment, no idea if we'll need it.
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>

Change-Id: I46a979cf2e9c0c22f3a799e530f63cf9d726756f
Reviewed-on: http://nxm.coreboot.org/review/158
Tested-by: Jenkins QA
Reviewed-by: John Floren <john@jfloren.net>