Grow the stack structure to 2 pages
The Base Of Stack (also wrongly called Top Of Stack) struct contains readonly
fields that should be protected, including the PID, but can be overwritten
at any point by an errant program.

This protection is becoming particularly important
as we chase a nasty Go problem. Further, it has proven very hard to grow the
struct over time because each growth requires moving it in memory, as it
is packed into an exactly-sized space. One can not change the struct
without recompiling all user code and libraries, losing the back-compatibility
for which Plan 9 is justly famed. In the current age, this kind of
hair-splitting makes no sense.

This change grows the starting stack to include one page at BOS which will
include the so-called "TOS" struct. On NXM, this means that USTKTOP
is now 4M, not 2M in size. BFD.

Later changes will involve write-protecting BOS. For a user to change
BOS they will need to write to a file (/proc/pid/bos?) which will protect
against programmer errors that trash the struct and are not caught until
some time later (e.g. the current problem that seems to be trashing g0)

Change-Id: If9e2403e1562d46d2bc0c0df60cf696f2187489f
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/159
Tested-by: Jenkins QA
Reviewed-by: Akshat Kumar <seed@mail.nanosouffle.net>