Add regression testing.
Creat a target. The target includes a file which is essentially a script.
Add kernel building to the standard mkfile, so we always build
kernels as part of jenkinshook.

The jenkinshook file is designed to be run from a script outside
this directory, but that script is simple, e.g.
export JENKINSHOOK=/home/rminnich
bash jenkinshook

Also it would be best if we could build a
disk image with grub and boot that in qemu.

The next step is to modify my jenkinshook sript to run
this jenkinshook; then the jenkins hook and regression
testing will be contained entirely in the NxM tree.

Change-Id: I4acaa3d7d4b2c856b841f89e9de86e7a2ef2a213
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/217
Tested-by: Jenkins QA
Reviewed-by: John Floren <john@jfloren.net>