Further progress on the jenkins hook + qemu
This now works just fine at nxm.coreboot.org, for both
negative and positive outcomes. Because of the
unsynced TSC problem we have to wire it to a core
using the taskset command.

Here is the command I use for qemu, which is in my
~/bin:

 ---
taskset 2 /home/rminnich/qemu-1.2.2/x86_64-softmmu/qemu-system-x86_64 -L /home/rminnich/qemu-1.2.2/pc-bios/ $*
 ---

Here is the script that runs this, it is in my login directory:

JENKINSHOME=/home/rminnich/ sh -x ./jenkinshook

Currently, this runs to completion and exits with 0 status
(i.e. succeeds). As far as I'm concerned this is ready
to be turned on.

Change-Id: If32495890418efc01f81dc56a70fc3d125f85e79
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/235
Tested-by: Jenkins QA