cppnsec: import from nix (accepted)
The change in nsec() not to cache fds broke cpp.
The problem is in cpp, not in nsec. Because it checks if
fd <= 0 and, somehow, it's called with 0 closed from the ape
framework. Now that we don't have hidden open fds, open returns 0
and ape thinks a file does not exist.
I think this fixes the problem.

Change-Id: I4046fe2739abd6fbcfea1cd6b2586996597d09db
Reference: /n/patches.lsub.org/patch/cppnsec
Signed-off-by: nemo@lsub.org
Signed-off-by: David du Colombier <0intro@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/250
Reviewed-by: ron minnich <rminnich@gmail.com>
Tested-by: Jenkins QA