mpvirtualbox: import from nix (accepted)
took out the change that was contended and which had nothing to do with the patch itself
I leave it for when there is agreement. This is only the mp tables change for virtualbox

Update 1: Added back the empty line before a function, somehow it got deleted in this
change.

Change-Id: Ib450e00b10b7c6e656eb19a87d86edf82b3adfa4
Reference: /n/patches.lsub.org/patch/mpvirtualbox
Signed-off-by: paurea@lsub.org
Signed-off-by: David du Colombier <0intro@gmail.com>
Signed-off-by: John Floren <john@jfloren.net>
Reviewed-on: http://nxm.coreboot.org/review/265
Tested-by: Jenkins QA