AHCI initial import
This is an early import of Erik's AHCI code. It seems to work on real hardware
but not in QEMU. It's *closer* to working with QEMU's Q35 chipset but not yet
there.

If some of us test this and it works ok on real hardware, this may be ready to
submit.

Edit: Removed libfis, since there's a separate change for it.

Change-Id: Ifd49b39a6e08e4c7d6ea7eb9f5d5b2f0856c7be8
Signed-off-by: John Floren <john@jfloren.net>
Reviewed-on: http://nxm.coreboot.org/review/270
Tested-by: Jenkins QA
Reviewed-by: ron minnich <rminnich@gmail.com>