Add a regression device.
This device is intended for the regression kernel, to allow us to
poke things we know to be troublesome. Normally the verbosity is zero,
as regression should cause the kernel to panic or hang, which the
regression harness will catch. If you want to see more verbosity,
echo -n V > /dev/regressctl
for less,
echo -n v > /dev/regressctl
it won't go below 0.

We currently only have one file, malloc. To use it, echo a size to it, e.g.
echo -n 0x2000000  >>/dev/malloc

This will cause a malloc and free. Currently, we use it to exercise a
known bug in the allocator. Fix is on the way after this patch.

Change-Id: I5a8f2ee39aa64b84c5ecf86eefb149c368f7394a
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/272
Tested-by: Jenkins QA
Reviewed-by: John Floren <john@jfloren.net>