nxm/k10: save some debugging space during notify
Save 128 bytes of space for debugging data between the
note frame and process stack pointer during note handling.
This is needed for Go in order to save the PC for trace-back
during a panic (caused by a synchronous signal). The Plan 9
kernel saves 256 bytes here, but our NFrame contains more
data and we shouldn't exceed 512 bytes for note handling.

Change-Id: Ia0a9d479d22ac6e81bb86fe3f6c7a6138cce4824
Signed-off-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-on: http://nxm.coreboot.org/review/309
Tested-by: Jenkins QA
Reviewed-by: ron minnich <rminnich@gmail.com>