libc/port: Add proper handling of subnormals, NaNs, and Infs to frexp
This fixes the behaviour in frexp(2) for NaN, Inf, and subnormal
number inputs: upon NaN or Inf, we return the respective input
and set the exponent to zero (instead of continuing with the
computation); subnormal numbers are first normalized so as to
make 20 ("SHIFT") bits of space for computing the appropriate
exponent (instead of always setting it to -1022).

A regression test has also been added to test for the extreme
cases: NaN, Inf, the smallest positive double, and the largest
negative double.

With this, we pass all of the tests in package "math" in Go.

Change-Id: Ie9c4110a7cbd2983844600222050e3640bf4bb05
Signed-off-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-on: http://nxm.coreboot.org/review/322
Tested-by: Jenkins QA
Reviewed-by: ron minnich <rminnich@gmail.com>