"Fix" mpacpi for AMD
On our test system, a reasonably recent AMD box, we enter mpacpi() with apics
== nil. But the function doesn't check for that; instead, it blithely goes
along de-referencing pointers all over the place. The outcome is unsurprising.

This patch just bails out immediately if the apics variable is not set.
Presumably it will continue to work for hardware that can use it, while we at
least avoid any degredation of performance.

Change-Id: I807d2f72218ad962f6eef92ab74ec36026304052
Signed-off-by: John Floren <john@jfloren.net>
Reviewed-on: http://nxm.coreboot.org/review/360
Tested-by: Jenkins QA
Reviewed-by: ron minnich <rminnich@gmail.com>