Implement a nanotime system call
It is getting increasingly difficult to reconcile what Go must have
with what we need to do in order to use /dev/bintime. The maintenance
of the fd's is turning into a clumsy, unreliable mess. In order to
simplify things, and improve higher accuracy for time, we add the
nanotime system call, which returns a vlong nanosecond count as
provided by todget(). We count on the ability to return 64 bits
directly as we don't want to do a validaddr() in the system call,
that just adds jitter.

This CL does nothing more than add the system call. Once this is
approved, a subsequent CL will include a replacement for nsec, to be
followed by a regression test.

Change-Id: I8a2e1f63e1193cd716285b6f2404e072dfbaad62
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/361
Tested-by: Jenkins QA
Reviewed-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-by: Jim McKie <borgled@gmail.com>