Implement an efficient gettimeofday for Linux
Lots of HPC apps really want to call gettimeofday, and it can turn
into a performance-busting problem. We can't reform every programmer
in the world. This is a likely correct stab at a real
gettimeofday, which we hope will improve miniFE performance.

You can't do file IO for everything, must as I wish we could.

john: fixed the sizes for the timeval struct, also tweaked how we fill it
and the debug output.

Change-Id: I2772feba6ff098aac5f4c990886a8043a858e14d
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Signed-off-by: John Floren <john@jfloren.net>
Reviewed-on: http://nxm.coreboot.org/review/369
Tested-by: Jenkins QA