Reinstate misplaced incref/decref optimisation.
Replace calls to incref/decref with a direct calls to either ainc/adec.
We originally did this optimisation on BG/P and applied it to AMD64 as well;
somehow it got lost in the journey to where we are now. For example,
profiling syspwrite, almost a third of the cycles (NOT including the actual
device write itself) spent in administration were in incref/decref.

It's not possible to immediately remove the Lock from the Ref structure,
there are cases where it is used directly as a lock on the containing
structure.

It's questionable whether to bracket the more portable incref/decref
implementations in arch.c with #ifdef/#endif or just remove them;
in this case they could be viewed as an aid to the code maintainer.

Change-Id: I25f83a23a7dfd837d3ae8ef302d16f5efdae3567
Signed-off-by: Jim McKie <borgled@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/370
Tested-by: Jenkins QA
Reviewed-by: Akshat Kumar <seed@mail.nanosouffle.net>
Reviewed-by: ron minnich <rminnich@gmail.com>