Fix regression test and correctoutput to not include a stack address
The badsyscall test included a print of the address of the fd array.
This changes as nxm changes. Get rid of that print and remove it from
correctoutput. This is what broke the nonixcalls patch, among other things.

Take the opportunity to fix a possible overflow problem in the test as well.
It was a problem as it happens. Nice catch Jim.

Change-Id: I792539c97ebfa10bcbf5927f322f13ff0640d461
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/371
Tested-by: Jenkins QA
Reviewed-by: Akshat Kumar <seed@mail.nanosouffle.net>