Store system call parameters in registers
Go system calls are really complex right now, because
our args are on the stack. Akshat has requested that we
get the system calls in registers. We already do this
for Linux, so it's not a big deal. We indicate the new
model by putting 0xc000 in AX, instead of 0x8000.

The next step is reworking how we save things in the assembly
level system call handling, followed by changes to
syscall(). This one is trick so we're taking it in steps.

Change-Id: I6b77b8cc5c8a921ea169bf0a9c33c01e3340737d
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/405
Tested-by: Jenkins QA
Reviewed-by: Akshat Kumar <seed@mail.nanosouffle.net>