Set up final preparations for register system calls
Change the function signature of syscall.
I originally wanted syscall(scallnr, Ureg *, uintptr []) which was
easy from .s but hard from tcore.c. So it now names all 5 args.

WHile we're at it, fix mksys.go to add the 5th arg.

All that remains is to start using the register args. That's
easy for the l64syscall.s part. The tcore.c code will be a tad
trickier but not by much. Also, the tcore.c code needs to be
adjusted to handle Linux system calls too.

Change-Id: I657f224f08a969512e76ee62fd552c105931ca89
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/409
Tested-by: Jenkins QA