Fix the AC syscall handling.
Somehow, in 2011, we managed to pick the wrong register for system
call #. Fix that, which will allow us to fix linux procs on ACs next.

As a side effect, because 60s is no longer quite enough, jack up the jenkinshook
time to 90s. Else the test for this commit may fail. It has once.

Change-Id: I4a314c5cc54af10cb546ef765faf640725aaf76a
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: http://nxm.coreboot.org/review/418
Tested-by: Jenkins QA
Reviewed-by: Akshat Kumar <seed@mail.nanosouffle.net>