More special cases for clang
We get further but now there is some weird isue with rc and mk... the arguemnts
are not seperated correctly in the for!

% mk all
for(i in 9sys 9syscall fmt port $objtype)@{
	echo $i
	cd $i
	mk $MKFLAGS install
}
9sys 9syscall fmt port
Can't cd 9sys 9syscall fmt port: No such file or directory
mk: for(i in 9sys ...  : exit status=exit(1)

Weird eh? the set of names is one arg. Any ideas?

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>