mksys utility from NxM.
this works and builds code that clang, at least, can assemble.

The little trick we do is put the system call into r9, the 6th argument.

This works very well, because then a system call is two instructions:
movq $syscallnumber, %r9
sysenter

Because we don't need to shuffle arguments around to fit the artificial
convention that the syscallno has to be the first argument. It does not.

This REQUIRES fixing the kernel system call code, but that can come later
when we can actually compile. Again, NxM can give us what we need.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>