Add a trivial Makefile to make emacs easier; add simple io funcs
I really dislike adding a Makefile but emacs keeps wanting it. Feel
free to delete if you don't like it.

I really dislike inline asm but this is known good, from coreboot.
I would rather we move this from .c to .S but I will bow to the
will of the people :-)

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>