We can now debug. And the klibc approach works.
And we're linking in bits of libc now.

There's a REALLY weird issue with fmtinit, where in the middle of setup we end
up in xactouser. I wonder if we're not taking an interrupt.

=> 0xfffffffff012afb0 <print+100>:
    mov    %rax,-0xb8(%rbp)
(gdb)
0xfffffffff012afb7 in print ()
1: x/i $pc
=> 0xfffffffff012afb7 <print+107>:
    mov    -0x1c8(%rbp),%rdx
(gdb)
0xfffffffff012afbe in print ()
1: x/i $pc
=> 0xfffffffff012afbe <print+114>:
    mov    -0xb8(%rbp),%rcx
(gdb)
0xfffffffff012afc5 in print ()
1: x/i $pc
=> 0xfffffffff012afc5 <print+121>:
    lea    -0x1c0(%rbp),%rax
(gdb)
0xfffffffff012afcc in print ()
1: x/i $pc
=> 0xfffffffff012afcc <print+128>:
    lea    0x100(%rax),%rsi
(gdb)
0xfffffffff012afd3 in print ()
1: x/i $pc
=> 0xfffffffff012afd3 <print+135>:
    lea    -0x1c0(%rbp),%rax
(gdb)
0xfffffffff012afda in print ()
1: x/i $pc
=> 0xfffffffff012afda <print+142>:
    mov    %rax,%rdi
(gdb)
0xfffffffff012afdd in print ()
1: x/i $pc
=> 0xfffffffff012afdd <print+145>:
    callq  0xfffffffff015ebf9 <xactouser>

(note how I'm using gdb :-)

no idea why this is happening.

Pleaase not hints in main.c on how to use gdb.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>