Implemented idtput.
I think we have to also get gdpput done correctly. What's there is totally
wrong; see vsvminit and how it calls gdtput to see what it needs.

We're taking an interrupt after spllo succeeds and we are not set up
correctly for them, which is why we explode on the retq from spllo.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>