Don't change core0 to KC
let's see if it can now timeshare.

The changes may be more far reaching.

If we can't get anywhere with this, I suggest we turn back the clock to the
July 2011 version of NIX. These AMP changes are just too far-reaching and
I doubt we care.I think we could recreate the kernel state pretty quickly.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>