Set mach on the bsp
The bsp has a special mach pointer, which we set in entry.S.
Since we want this set very early, set it as the first thing we do in main.

Now, we can use machp() everywhere (we hope) which will make SMP possible
(we hope) sooner than we thought (we hope).

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>