Begin removing references to the global Mach variable, m
We don't have the luxury of using m because we don't have kenc any more.

To get smp to work we have to get rid of m (and then up)

Doing it all at once has never worked; so we're going to take this one file or a few
files at a time.

Taslock.c is an early canary in the coal mine, since it's use early. If taslock.c works,
the signs are good. In this case, it works.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>