Moving from up to m->externup
This was the path recommended by Charles Forsyth: put and externalup pointer
in Mach, and use m->externalup.

Note the little places we put
Mach *m = machp();
at the start of functions using m.

There's another painful round of changes to do where we get rid of the global
Mach *m;
which used to be R15, and move it all to machp().

This fails at some point but we have to do it. The use of up is just going to kill us
when we go to SMP. And it's not even right now.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>