Remove uses of the global Mach variable in port/
In preparation for removing the global Mach variable, remove uses in port.
Removal in k10/ is a bit trickier but it's coming. This change is required
if we ever want more than one core, since we no longer have kenc and can't
use R15 for the per-core Mach pointer, and it's no longer tied to the the

extern register Mach *m;

which looked like a global but was not; instead, it just used one of two
dedicated registers to hold the per-core Mach pointer.

We have to some sort of per-core register so we use %gs, and to read it, we have
call machp(), since we have no way to connect a "global" Mach *m to %gs in
gcc (of which we are aware).

The tool I used was coccinnelle, which automates this nonsense.
The script looks like this:

@r exists@
typedef Mach; // only needed once per semantic patch
idexpression Mach *m;
function f;
@@
f(...){
<+...
m
...+>
}

@@
function r.f;
@@

f(...) {
+ Mach *m = machp();
...
}

The command is
spatch --sp-file mach.cocci -dir port --in-place

Thanks to Julia Lawall for pointing the way to the working script.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>