Add the "poor mans ftrace" from Akaros.
Typical usage:

strace --sp-file sys/src/spatchs/poor-ftrace.cocci port/dev.c --in-place

note that when you run you get a nice trade of function calls in dev.c.

I'm leaving this in dev.c for now for you to see. It shows that we're calling devwalk a lot,
which is pretty interesting ...

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>