We're calling a system call from init. But the args are set up wrong for syscall()
with the result that our uregs pointer is utterly bogus.

What we should do: set up uregs as we do in interrupts, call syscall with just the uregs
pointer, and pull the syscallno and args out of uregs. Much simpler.

Or, assume callee save works, and don't save anything at all, since syscall (it's a function) should
just save everything.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>