Remove the global mach declaration from dat.h; add them to each file.
The program in util (removemach.go) did this.

It looks at a .o for an undefined symbol named 'm'. If it finds
it, it finds the source file and rewrites the file to have the
declaration at the head, along with the words REMOVE ME.

These words can then identify files to be processed by
coccinnelle.

This seems to have worked, happily. We're on our way to SMP.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>