Remove file scope use of Mach *m in port and k10
At this point, all global use of m is gone, which means we
we will be ready for SMP when the time comes. All references
to m are now
m = machp()
and machp() returns
move %gs, %rax
i.e. the m pointer *for that core*. Further, the vsvminit code,
for each core, properly sets the GSBase MSR for that core.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>