Debugging snapshot for sysrforkret showing a few techniques
sysrforkret is not quite right. With this commit I'm showing
a few techniques you can use to zero in on the problem. Of course,
this assumes you know how to attach with gdb and take it from
there!

FYI, the RETQ is failing, badly, so that's a start. It looks
like it's using the value of ds (0x2b) as the return IP? The
stack is wrong.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>