Add counters, incremented by assembly code, for interrupts and system calls
These counters show a way you can set 'tracking' visible in C and gdb.
Basically, we define the counter in trap.c, but increment it in
assembly. Then, in gdb, we can see the values. Or we can print them.

If ire is > irx, we're in an interrupt.
if sce > scx, we're in a system call.
lastvno shows the previous vno of the one we're taking now.
This might be useful; if lastvno is always the same it points to a bug
in a driver.

In the current bug where we're taking a double fault, at the time we die,
sce is 82, and scx is 81
ire is 15220, and irx is 15219, we're in an interrupt.

I feel like the system call is getting interrupted at the wrong time, OR
some piece of interrupt handling code is trashing the stack.

What we can do is add more counters, and get a finer grain idea of where we
are when it goes bad. Also, if the ir* and sc* values are always the same,
this is so deterministic it makes life much easier.

ire and irx change, sce and scx are the same. Race condition? ah joy.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>