Break everything. We have to fix this however.
This change sets the uart to polling and as a result
breaks everything completely. But we need to fix this now.

As you can see, you won't get very far. Which is good, as we've
been living in a fool's paradise.

My *guess* is that we're taking an interrupt in the middle
of a system call and it's not ending well.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>