Sample instrumentation to try to catch what is going on
This is just an example of what you might do to find it.

I would not trust anything -- do our locks work? Are we
allocating the same page to two different people? You might
want to write a littel FIFO code to log the allocations
of each page as they happen, not printing, but to a fifo
in memory you can dump in gdb.

Just think about things you can do that don't involve printing,
because this is a race condition and printing will make it
disappear.

Anyway, see what you think. In gdb, you can do things like
print km
print npc (new page counter)
and so on.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>