Prototype of the build tool.
This gets surprisingly far, and it's really fast.

Note that in the final version, the BUILD.json will be in something
like k8cpu.json, and will include core.json, i.e. stuff you have to have.
You can name the top level file, and it might default to BUILD.json
if it's a directory.

There will need to be command directives to add things to the rootfs and so on,
but I doubt that will be hard. See what you think.

I think as a first step we can start to replace parts of the top level
BUILD script with invocations of this tool, and the use of this tool will grow
over time and the amount of stuff done in the BUILD script will shrink, until the BUILD
script is gone entirely.

The idea of the name BUILD.json is you will be able to specify a directory, and it will
walk the directory and fine the various BUILD files and do what needs doing.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>