Bread out various .json bits into other files.
It's unlikely that we ever need to touch
core, ip, port, .json.

Just the individual elements from, e.g., k8cpu would have to change.
Still need to get special rules and such in.

The Project element is in case you need to do a clean build of a totally
different set of things, e.g. init9.c or similar. It starts with a clean
build struct; it's like a totally seperation invocation of build.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>