We have a kernel build tool
Do this:
cd util && go build build.go
PATH=$PATH:/path/to/harvey/util

cd sys/src/9/k10
/path/to/harvey/util/build k8cpu

gets you a bootable kernel.

Next step is to add .json files to build the tools we put into the root,
but that should be simple.

This is not the end, of course, but I'm trying to show what's possible with a go build tool.
We can make it simple and fast and reliable. And it's more compact and reliable than shell scripts.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>