Add the lindent ("Linux indent") script
I think we should require, as much as possible, that all C code follow
the linux kernel standard. Use this script on your C code.

I don't intend to express approval of Linux kernel coding style, just that it is a standard, we follow it on many projects, and
this script makes it easy to follow.

This is my attempt to have us not argue about C style. Just pick a standard, use it, and hold your nose if you have to :-)

Change-Id: I06a18407510e31850e690883cde483e40bb1857a
Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>